Don't hard-code runtime application path
It is not guaranteed that the app is installed to /app. Users of buildpacks other than Heroku require applications to be stored at other paths, such as [pkgr](http://github.com/crohr/pkgr) which uses /opt/${APP_NAME}. Furthermore, the source may not be located at /app at compile time, since not every buildpack executor uses a layered filesystem, and so must resort to using a sandbox build directory (some call this the good old days).

See https://github.com/crohr/pkgr/issues/93 for the source of this.